src/api/config/__init__.py
from dotenv import find_dotenv, load_dotenv

if not find_dotenv():
    exit("Cannot load environment variables. There's no .env file")
else:
    load_dotenv()


from . import apidocs, aws, broker, hash_cache, metadata_cache, postgres, text_cache

__all__ = [postgres, aws, hash_cache, metadata_cache, text_cache, broker, apidocs]
