pyproject.toml
[tool.poetry]
name = "pet-pastebin"
version = "0.1.0"
description = "pastebin web app helps users to share huge text blocks by short URLs"
authors = ["Griy.py"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"


[tool.poetry.group.api.dependencies]
apispec = "^6.6.1"
apispec-webframeworks = "^1.1.0"
boto3 = "^1.34.121"
boto3-stubs = {extras = ["s3"], version = "^1.34.121"}
celery = "^5.4.0"
flasgger = "^0.9.7.1"
flask = "^3.0.3"
flask-restful = "^0.3.10"
gunicorn = "^22.0.0"
marshmallow = "^3.21.3"
mypy-boto3 = "^1.34.121"
psycopg2-binary = "^2.9.9"
redis = "^5.0.5"
sqlalchemy = "^2.0.30"


[tool.poetry.group.gen.dependencies]
psycopg2-binary = "^2.9.9"
redis = "^5.0.5"
sqlalchemy = "^2.0.30"


[tool.poetry.group.queue.dependencies]
boto3 = "^1.34.121"
boto3-stubs = {extras = ["s3"], version = "^1.34.121"}
celery = "^5.4.0"
mypy-boto3 = "^1.34.121"
psycopg2-binary = "^2.9.9"
redis = "^5.0.5"
sqlalchemy = "^2.0.30"


[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
requests = "^2.32.3"


[tool.poetry.group.frontend.dependencies]
aiohttp = "^3.9.5"
fastapi = "^0.111.0"
fastui = "^0.6.0"
uvicorn = "^0.30.1"



[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
